requirements.txt
pyTelegramBotAPI
psutil
google-generativeai
Pillow
pyautogui
opencv-python